digikam from trunk:
The ImageEditor can now access the database independently
from main view, i.e. also if main moved to a different directory.

- adjust ImageInfo to be more like a database item wrapper class
- add a method to ImageWindow to start it with a list of
  ImageInfo objects, use this in AlbumIconView
- adapt ImageWindow to (optional) ImageInfo list
- adjust behavior of saveAsIsComplete
- add itemChanged methods to ImagePropertiesSidebarDB:
  - one takes only a URL (no database)
  - one takes an ImageInfo object (used by ImageWindow, full DB support)
  - one takes AlbumIconItem and -View (used by main view,
    full DB support, optional future interaction with the view)
- ImageDescEditTab now works on an ImageInfo object
- add a copyItem method to ImageInfo as a wrapper for said function from AlbumDB
- modify AlbumDB::copyItem to return image id of inserted image
- add DISTINCT keyword to second SELECT statement of AlbumDB::copyItem.
  With all of above changes, I constantly produced bug #113915.
  I do not know why the SELECT statement produces (in my case) three times
  the same result so that INSERT complains, and I do not know which of
  my changes should ever trigger this problem, but the DISTINCT fixes this.
  If anyone speaks SQL, you might want to check my comment in AlbumDB::copyItem.

Currently, the rating is not copied when database information is copied (in copyItem).
Should this be done?

CCBUG: 113915
CCMAIL: digikam-devel@kde.org


svn path=/trunk/extragear/graphics/digikam/; revision=529254

+++ digikam/albumdb.cpp
@@ -1173,9 +1173,13 @@ void AlbumDB::moveItem(int srcAlbumID, const QString& srcName,
                   QString::number(srcAlbumID), escapeString(srcName)) );
 }
 
-void AlbumDB::copyItem(int srcAlbumID, const QString& srcName,
-                       int dstAlbumID, const QString& dstName)
+int AlbumDB::copyItem(int srcAlbumID, const QString& srcName,
+                      int dstAlbumID, const QString& dstName)
 {
+    // check for src == dest
+    if (srcAlbumID == dstAlbumID && srcName == dstName)
+        return -1;
+
     // first delete any stale database entries if any
     deleteItem(dstAlbumID, dstName);
 
@@ -1185,15 +1189,26 @@ void AlbumDB::copyItem(int srcAlbumID, const QString& srcName,
              .arg(QString::number(dstAlbumID), escapeString(dstName),
                   QString::number(srcAlbumID), escapeString(srcName)) );
 
+    int id = sqlite3_last_insert_rowid(m_db);
+
+    // I think this can be done much easier:
+    /*
     execSql( QString("INSERT INTO ImageTags (imageid, tagid) \n"
-                     "SELECT I.id, T.tagid FROM Images AS I, ImageTags AS T WHERE \n"
+                     "SELECT %1, tagid FROM ImageTags \n"
+                     "WHERE imageid=%2 ;)
+              .arg(QString::number(id), QString::number(id)) );
+    */
+
+    execSql( QString("INSERT INTO ImageTags (imageid, tagid) \n"
+                     "SELECT DISTINCT I.id, T.tagid FROM Images AS I, ImageTags AS T WHERE \n"
                      "     I.id=(SELECT Images.id FROM Images WHERE \n"
                      "           dirid=%1 AND name='%2') \n"
                      "AND  T.tagid IN (SELECT tagid FROM ImageTags WHERE \n"
                      "                   imageid=(SELECT id FROM Images WHERE \n"
-                     "                             dirid=%3 AND name='%4'))")
+                     "                             dirid=%3 AND name='%4'));")
              .arg(QString::number(dstAlbumID), escapeString(dstName), 
                   QString::number(srcAlbumID), escapeString(srcName)) );
+    return id;
 }
 
 Q_LLONG AlbumDB::lastInsertedRow()
